combined host and lib files to single file extensions
