+++ src/lib.rs
@@ -1,13 +1,9 @@
 mod cgi;
-mod cgi_host;
 mod error;
 mod http;
-mod http_host;
 mod llm;
 mod memory;
-mod memory_host;
 mod socket;
-mod socket_host;
 
 pub use cgi::*;
 pub use error::*;
