+++ src/memory.rs
@@ -1,4 +1,11 @@
-use crate::memory_host::*;
+#[link(wasm_import_module = "blockless_memory")]
+extern "C" {
+    #[link_name = "memory_read"]
+    pub(crate) fn memory_read(buf: *mut u8, len: u32, num: *mut u32) -> u32;
+    #[link_name = "env_var_read"]
+    pub(crate) fn env_var_read(buf: *mut u8, len: u32, num: *mut u32) -> u32;
+}
+
 
 pub fn read_stdin(buf: &mut [u8]) -> std::io::Result<u32> {
     let mut len = 0;
